requirements.txt
argparse
tqdm
jieba
random
pypinyi